+++ packages/react-craft-ai-decision-tree/stories/ParentComponent.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable react/jsx-no-bind */
 import DecisionTreeWithPanel from '../src';
 import tree from './tree.json';
 import React, { useState } from 'react';
@@ -14,8 +15,10 @@ const ParentComponent = () => {
   return (
     <div>
       <DecisionTreeWithPanel
-        width={ 600 }
-        height={ 400 }
+        style={{
+          width: 600,
+          height: 400
+        }}
         data={ tree }
         updatePositionAndZoom={ updatePosAndZoom }
         scale={ savedNewZoom }